+++ src/main/java/edu/rodriguezestrada/laboratoriomodysim/simulacion/probabilidad/Probabilidad.java
@@ -6,5 +6,5 @@ package edu.rodriguezestrada.laboratoriomodysim.simulacion.probabilidad;
  * @author gestrada
  */
 public interface Probabilidad {
-    public Integer generarValor();
+    public Double generarValor();
 }
